OOP - Code reuse with classes
